update code products.js, finish cart and build form

+++ front/js/cart.js
@@ -1,16 +1,14 @@
 // Auto appel des fonctions
 (async function() {
-    // let productId = getProductId()
-    let product = await getProducts()
+    const product = await getProducts()
     displayCart()
     updateQuantityEvent()
     removeElementEvent()
     totalQuantity()
     price(product)
-    // totalPrice()
-    form()
+    totalPrice()
+    SaveForm()
 })()
-
 // recup des produits avec l'API
 function getProducts() {
     return fetch("http://localhost:3000/api/products/") 
@@ -22,16 +20,14 @@ function getProducts() {
 function getStorageProduct() {
     return JSON.parse(localStorage.getItem("products"))
 }
-
 // Gestion panier
 function displayCart() {
-
     // Action si panier vide ou rempli
     if(getStorageProduct() == null) {
         document.querySelector("h1").innerText = "Votre panier est vide"
         document.querySelector(".cart__price").innerHTML = ""
     } else {
-        let displayProduct = document.getElementById("cart__items")
+        const displayProduct = document.getElementById("cart__items")
         getStorageProduct().find(element => {
             displayProduct.innerHTML += `
             <article class="cart__item" data-id="${element.id}" data-color="${element.color}">
@@ -42,7 +38,7 @@ function displayCart() {
                 <div class="cart__item__content__description">
                     <h2>${element.name}</h2>
                     <p>${element.color}</p>
-                    <p id="priceProduct"> €</p>
+                    <p id="priceProduct"></p>
                 </div>
                 <div class="cart__item__content__settings">
                     <div class="cart__item__content__settings__quantity">
@@ -58,27 +54,23 @@ function displayCart() {
         })
     }
 }
-// Ajout Prix de l'API
-// function price(product) {
-//     let getPriceDom = document.querySelectorAll("#priceProduct");
-
-//     getPriceDom.forEach((element, i) => {
-//         product.forEach((item, j) => {
-//             let getIdProduct = product[j]._id
-//             let getIdStorage = getStorageProduct()[i].Id
-//             if (getIdStorage === getIdProduct) {
-//                 getPriceDom == product[j].price
-//             }
-//         })
-//     })
-// }
+// Ajout Prix de l'API avec calcul individuel par quantité
+function price(product) {
+    const priceProduct = document.querySelectorAll("#priceProduct");
+    const inputQuantity = document.querySelectorAll(".itemQuantity");
+    priceProduct.forEach((item, i)  => {
+        const id = item.closest(".cart__item").dataset.id
+        const find = product.find(element => element._id == id)
+        item.innerText = (find.price * inputQuantity[i].value)+ " €"
+    })
+}
 // changement des quantités avec update du localStorage
 function updateQuantityEvent() {
-    let inputQuantity = document.querySelectorAll(".itemQuantity");
-    let storage = JSON.parse(localStorage.getItem("products"));
-        storage.forEach((eltItem, i) => {
+    const inputQuantity = document.querySelectorAll(".itemQuantity");
+    const storage = JSON.parse(localStorage.getItem("products"));
+        storage.forEach((product, i) => {
         inputQuantity[i].addEventListener("change",() => {
-            let objIndex = storage.findIndex(item=> item.id === eltItem.id && item.color === eltItem.color)
+            const objIndex = storage.findIndex(item=> item.id === product.id && item.color === product.color)
             if (objIndex !== -1) {
                 storage[objIndex].quantity = inputQuantity[i].value
             }
@@ -89,31 +81,48 @@ function updateQuantityEvent() {
 }
 // Fonction suppression du produit dans le panier
 function removeElementEvent() {
-    let deleteBtn = document.querySelectorAll(".deleteItem");
-    let storage = JSON.parse(localStorage.getItem("products"));
+    const deleteBtn = document.querySelectorAll(".deleteItem");
+    const storage = JSON.parse(localStorage.getItem("products"));
     deleteBtn.forEach((element, i) => {
         deleteBtn[i].addEventListener("click", () => {
-            let idRemove = storage[i].id
-            let colorDelete = storage[i].color
-            let updateRemoveStorage = storage.filter((product => product.id !== idRemove || product.color !== colorDelete))
-            localStorage.setItem("products", JSON.stringify(updateRemoveStorage)); 
-            location.reload()
+            if (window.confirm("Voulez-vous vraiment supprimer le produit sélectionné?")) {
+                const idRemove = storage[i].id
+                const colorDelete = storage[i].color
+                const updateRemoveStorage = storage.filter((product => product.id !== idRemove || product.color !== colorDelete))
+                localStorage.setItem("products", JSON.stringify(updateRemoveStorage)); 
+                location.reload()
+            }
         })
     })
+    if (deleteBtn.length == 0 ) {
+        localStorage.removeItem("products")
+        location.reload()
+    }
 }
 // Calcul et affichage des quantités
 function totalQuantity() {
-    let totalQuantity = document.getElementById("totalQuantity");
-    let quantityProduct = document.querySelectorAll(".itemQuantity");
+    const totalQuantity = document.getElementById("totalQuantity")
+    const quantityProduct = document.querySelectorAll(".itemQuantity")
     let totalNumber = 0;
     quantityProduct.forEach(quantity => {
         totalNumber += parseInt(quantity.value)
     })
     return totalQuantity.innerText = totalNumber
 }
-// Recup bouton "Commander!" et envoi dans le localStorage des valeurs remplies
-function form() {
-    let btnCommand = document.getElementById("order");
+// Calcul du prix total
+function totalPrice() {
+    const getTotalPrice = document.getElementById("totalPrice")
+    const priceProduct = document.querySelectorAll("#priceProduct")
+    let totalNumber = 0
+    priceProduct.forEach(element => {
+        const totalPrice = element.textContent
+        totalNumber += parseInt(totalPrice)
+        getTotalPrice.innerText = totalNumber
+    });
+}
+// Recup bouton "Commander!" et envoi dans le localStorage des valeurs remplies du formulaire
+function SaveForm() {
+    const btnCommand = document.getElementById("order");
     btnCommand.addEventListener("click", (e) => {
         e.preventDefault()
         let formValue = {
@@ -125,4 +134,20 @@ function form() {
         }
         localStorage.setItem("form", JSON.stringify(formValue))
     })
+
+}
+// Gestion validation du formulaire
+function validateForm() {
+    // recup messages erreur
+    const firstNameErrorMsg = document.getProductId("firstNameErrorMsg")
+    const lastNameErrorMsg = document.getProductId("lastNameErrorMsg")
+    const addressErrorMsg = document.getProductId("addressErrorMsg")
+    const cityErrorMsg = document.getProductId("cityErrorMsg")
+    const emailErrorMsg = document.getProductId("emailErrorMsg")
+    // regex
+    const regexFirstNameErrorMsg = new RegExp(/^[a-zA-Z\s'.-]+$/)
+    const regexaddressErrorMsg = new RegExp(/^[A-Za-z0-9'\.\-\s\,]+$/)
+    const regexEmailErrorMsg = new RegExp(
+        /^((\w[^\W]+)[\.\-]?){1,}\@(([0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3})|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+      )
 }
